Move wing and backstage specific code to the classes that use them
